requirements.txt
requests
python-dotenv
flask
google-cloud-firestore
google-auth